module: merge config with `package_json_reader`
PR-URL: https://github.com/nodejs/node/pull/50322
