+++ lib/internal/modules/esm/get_format.js
@@ -19,7 +19,7 @@ const {
 const experimentalNetworkImports =
   getOptionValue('--experimental-network-imports');
 const { containsModuleSyntax } = internalBinding('contextify');
-const { getPackageType } = require('internal/modules/esm/package_config');
+const { getPackageType } = require('internal/modules/package_json_reader');
 const { fileURLToPath } = require('internal/url');
 const { ERR_UNKNOWN_FILE_EXTENSION } = require('internal/errors').codes;
