+++ lib/internal/modules/esm/module_job.js
@@ -232,7 +232,7 @@ class ModuleJob {
         const packageConfig =
           StringPrototypeStartsWith(this.module.url, 'file://') &&
             RegExpPrototypeExec(/\.js(\?[^#]*)?(#.*)?$/, this.module.url) !== null &&
-            require('internal/modules/esm/package_config')
+            require('internal/modules/package_json_reader')
               .getPackageScopeConfig(this.module.url);
         if (packageConfig.type === 'module') {
           e.message +=
