+++ lib/internal/modules/esm/resolve.js
@@ -56,7 +56,6 @@ const {
 } = require('internal/errors').codes;
 
 const { Module: CJSModule } = require('internal/modules/cjs/loader');
-const { getPackageScopeConfig } = require('internal/modules/esm/package_config');
 const { getConditionsSet } = require('internal/modules/esm/utils');
 const packageJsonReader = require('internal/modules/package_json_reader');
 const { internalModuleStat } = internalBinding('fs');
@@ -688,7 +687,7 @@ function packageImportsResolve(name, base, conditions) {
     throw new ERR_INVALID_MODULE_SPECIFIER(name, reason, fileURLToPath(base));
   }
   let packageJSONUrl;
-  const packageConfig = getPackageScopeConfig(base);
+  const packageConfig = packageJsonReader.getPackageScopeConfig(base);
   if (packageConfig.exists) {
     packageJSONUrl = pathToFileURL(packageConfig.pjsonPath);
     const imports = packageConfig.imports;
@@ -797,7 +796,7 @@ function packageResolve(specifier, base, conditions) {
     parsePackageName(specifier, base);
 
   // ResolveSelf
-  const packageConfig = getPackageScopeConfig(base);
+  const packageConfig = packageJsonReader.getPackageScopeConfig(base);
   if (packageConfig.exists) {
     const packageJSONUrl = pathToFileURL(packageConfig.pjsonPath);
     if (packageConfig.exports != null && packageConfig.name === packageName) {
