+++ lib/internal/modules/package_json_reader.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const {
+  ArrayIsArray,
   JSONParse,
   StringPrototypeSlice,
   StringPrototypeLastIndexOf,
@@ -149,11 +150,42 @@ function getNearestParentPackageJSON(checkPath) {
   return { data, path };
 }
 
+/**
+ * Returns the package configuration for the given resolved URL.
+ * @param {URL | string} resolved - The resolved URL.
+ * @returns {import('typings/internalBinding/modules').PackageConfig} - The package configuration.
+ */
+function getPackageScopeConfig(resolved) {
+  const result = modulesBinding.getPackageScopeConfig(`${resolved}`);
+
+  if (ArrayIsArray(result)) {
+    return deserializePackageJSON(`${resolved}`, result, false /* checkIntegrity */);
+  }
+
+  // This means that the response is a string
+  // and it is the path to the package.json file
+  return {
+    __proto__: null,
+    pjsonPath: result,
+    exists: false,
+    type: 'none',
+  };
+}
+
+/**
+ * Returns the package type for a given URL.
+ * @param {URL} url - The URL to get the package type for.
+ */
+function getPackageType(url) {
+  // TODO(@anonrig): Write a C++ function that returns only "type".
+  return getPackageScopeConfig(url).type;
+}
+
 module.exports = {
   checkPackageJSONIntegrity,
   read,
   readPackage,
   getNearestParentPackageJSON,
-
-  deserializePackageJSON,
+  getPackageScopeConfig,
+  getPackageType,
 };
